requirements.txt
snowflake-connector-python>=3.6.0
pyyaml>=6.0
pytest>=8.0.0
